+++ CHANGES.rst
@@ -6,12 +6,17 @@ mqttwarn changelog
 in progress
 ===========
 
+
+2023-05-15 0.34.1
+=================
+
 - Add example "Simple MQTT media player". Thanks, @nagyrobi.
 - Add example "Forwarding data from IoT devices to Zabbix"
 - Remove support for Secure Sockets Layer Version 3.0 (SSLv3).
   As per RFC 7568, SSLv3 has been deprecated in 2015 already.
 - Tests: Add more test cases to increase mqttwarn core coverage to ~100%
 - Improve example "Forward OwnTracks low-battery warnings to ntfy"
+- Migrate documentation to https://mqttwarn.readthedocs.io/
 
 
 2023-04-28 0.34.0